main: consume messenger events and initialize with online peers
To start up our OnionMessenger with the correct starting state, we
call ListPeers on startup and produce PeerOnline events so that the
set of peers will be known to the messenger.

Note that we can't consume messenger events in a thread (it is
blocking) because it uses a RefCell to provide interior mutability (the
NodeSigner trait is immutable, but the LND signer client that it relies
on requires mutability).